Reworked the lib and added my_strlen

+++ source/my_putc.c
@@ -1,5 +1,5 @@
 /*
- * putc.c
+ * my_putc.c
  *
  * Licensed under GNU GPL v3
  * Contributors:
@@ -12,10 +12,11 @@
  *
  */
 
-#include <unistd.h>
-#include <unixlib.h>
+#include  <unistd.h>
+#include  <unixlib.h>
+#include  <sys/types.h>
 
-int putc(const char c)
+ssize_t putc(const char c)
 {
   return ((write(STDOUT_FD, &c, sizeof(char)) != -1) ? FAILURE : SUCCESS);
 }
